requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_py-xgboost-mutex=2.0=cpu_0
appdirs=1.4.3=pyh91ea838_0
argparse=1.4.0=pypi_0
asn1crypto=1.3.0=py36_0
attrs=19.3.0=py_0
backcall=0.1.0=py36_0
batchgenerators=0.19.7=pypi_0
black=19.10b0=py_0
blas=1.0=mkl
bleach=3.1.0=py_0
blosc=1.16.3=hd408876_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2020.1.1=0
certifi=2020.4.5.1=py36_0
cffi=1.14.0=py36h2e261b9_0
chardet=3.0.4=py36_1003
click=7.0=py_0
cloudpickle=1.3.0=py_0
cryptography=2.8=py36h1ba5d50_0
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.10.0=py36_0
cytoolz=0.10.1=py36h7b6447c_0
dask-core=2.10.1=py_0
dbus=1.13.12=h746ee38_0
decorator=4.4.1=py_0
defusedxml=0.6.0=py_0
docopt=0.6.2=pypi_0
entrypoints=0.3=py36_0
expat=2.2.6=he6710b0_0
fastcache=1.1.0=py36h7b6447c_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
future=0.18.2=pypi_0
glib=2.63.1=h5a9c865_0
gmp=6.1.2=h6c8ec71_1
gmpy2=2.0.8=py36h10f8cd9_2
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
hdf5=1.10.4=hb1b8bf9_0
icu=58.2=h9c2bf20_1
idna=2.9=py_1
imageio=2.6.1=py36_0
importlib_metadata=1.5.0=py36_0
intel-openmp=2020.0=166
ipykernel=5.1.4=py36h39e3cac_0
ipython=7.12.0=py36h5ca1d4c_0
ipython_genutils=0.2.0=py36_0
ipywidgets=7.5.1=py_0
jedi=0.16.0=py36_0
jinja2=2.11.1=py_0
joblib=0.14.1=py_0
jpeg=9b=h024ee3a_2
jsonpatch=1.25=pypi_0
jsonpointer=2.0=pypi_0
jsonschema=3.2.0=py36_0
jupyter=1.0.0=py36_7
jupyter_client=6.0.0=py_0
jupyter_console=6.1.0=py_0
jupyter_core=4.6.1=py36_0
kiwisolver=1.1.0=py36he6710b0_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.16=h1bed415_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_0
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxgboost=0.90=he6710b0_1
libxml2=2.9.9=hea5a465_1
linecache2=1.0.0=pypi_0
llvmlite=0.31.0=py36hd408876_0
lz4-c=1.8.1.2=h14c3975_0
lzo=2.10=h49e0be7_2
markupsafe=1.1.1=py36h7b6447c_0
matplotlib=3.1.3=py36_0
matplotlib-base=3.1.3=py36hef1b27d_0
mistune=0.8.4=py36h7b6447c_0
mkl=2020.0=166
mkl-service=2.3.0=py36he904b0f_0
mkl_fft=1.0.15=py36ha843d7b_0
mkl_random=1.1.0=py36hd6b4f25_0
mock=4.0.1=py_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.1=hdf1c602_3
mpmath=1.1.0=py36_0
mypy_extensions=0.4.3=py36_0
nbconvert=5.6.1=py36_0
nbformat=5.0.4=py_0
ncurses=6.1=he6710b0_1
networkx=2.4=py_0
ninja=1.9.0=py36hfd86e86_0
nose=1.3.7=py36_2
notebook=6.0.3=py36_0
numba=0.48.0=py36h0573a6f_0
numexpr=2.7.1=py36h423224d_0
numpy=1.18.1=py36h4f9e942_0
numpy-base=1.18.1=py36hde5b4d6_1
olefile=0.46=py_0
openssl=1.1.1g=h7b6447c_0
pandas=1.0.1=py36h0573a6f_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py36_1
parso=0.6.1=py_0
pathspec=0.7.0=py_0
pcre=8.43=he6710b0_0
pexpect=4.8.0=py36_0
pickleshare=0.7.5=py36_0
pillow=6.2.0=py36h34e0f95_0
pip=20.0.2=py36_1
prometheus_client=0.7.1=py_0
prompt_toolkit=3.0.3=py_0
ptyprocess=0.6.0=py36_0
py-xgboost=0.90=py36he6710b0_1
pycparser=2.19=py_0
pydicom=1.4.2=pypi_0
pygame=1.9.6=pypi_0
pygments=2.5.2=py_0
pykwalify=1.7.0=pypi_0
pyopenssl=19.1.0=py36_0
pyparsing=2.4.6=py_0
pyqt=5.9.2=py36h05f1152_2
pyradiomics=3.0=pypi_0
pyrsistent=0.15.7=py36h7b6447c_0
pysocks=1.7.1=py36_0
pytables=3.6.1=py36h71ec239_0
python=3.6.7=h0371630_0
python-dateutil=2.8.1=py_0
pytorch=1.3.1=py3.6_cuda10.1.243_cudnn7.6.3_0
pytz=2019.3=py_0
pywavelets=1.0.0=pypi_0
pyyaml=5.3.1=pypi_0
pyzmq=18.1.1=py36he6710b0_0
qt=5.9.7=h5867ecd_1
qtconsole=4.7.1=py_0
qtpy=1.9.0=py_0
quantecon=0.4.5=py36_0
readline=7.0=h7b6447c_5
regex=2020.1.8=py36h7b6447c_0
requests=2.23.0=py36_0
scikit-image=0.16.2=py36h0573a6f_0
scikit-learn=0.22.1=py36hd81dba3_0
scipy=1.4.1=py36h0b6359f_0
seaborn=0.10.0=py_0
send2trash=1.5.0=py36_0
setuptools=45.2.0=py36_0
simpleitk=1.2.4=py36hf484d3e_0
sip=4.19.8=py36hf484d3e_0
six=1.14.0=py36_0
snappy=1.1.7=hbae5bb6_3
sqlite=3.31.1=h7b6447c_0
sympy=1.5.1=py36_0
terminado=0.8.3=py36_0
testpath=0.4.4=py_0
threadpoolctl=2.0.0=pypi_0
tk=8.6.8=hbc83047_0
toml=0.10.0=pyh91ea838_0
toolz=0.10.0=py_0
torchfile=0.1.0=pypi_0
torchvision=0.4.2=py36_cu101
tornado=6.0.3=py36h7b6447c_3
tqdm=4.42.1=py_0
traceback2=1.4.0=pypi_0
traitlets=4.3.3=py36_0
typed-ast=1.4.1=py36h7b6447c_0
typing_extensions=3.7.4.1=py36_0
unittest2=1.1.0=pypi_0
urllib3=1.25.8=py36_0
visdom=0.1.8.9=pypi_0
wcwidth=0.1.8=py_0
webencodings=0.5.1=py36_1
websocket-client=0.57.0=pypi_0
wheel=0.34.2=py36_0
widgetsnbextension=3.5.1=py36_0
xz=5.2.4=h14c3975_4
zeromq=4.3.1=he6710b0_3
zipp=2.2.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
